Updates from review comments
