+++ src/SourceBuild/content/test/Microsoft.DotNet.SourceBuild.SmokeTests/OmniSharpTests.cs
@@ -21,8 +21,7 @@ public class OmniSharpTests : SmokeTests
 
     public OmniSharpTests(ITestOutputHelper outputHelper) : base(outputHelper) { }
 
-    // Disable the test when run offline because it requires downloading the OmniSharp tarball.
-    [SkippableTheory(new string[] { Config.ExcludeOmniSharpEnv, Config.OfflineEnv }, skipOnTrue: true)]
+    [SkippableTheory(Config.ExcludeOmniSharpEnv, skipOnTrue: true)]
     [InlineData(DotNetTemplate.BlazorWasm)]
     [InlineData(DotNetTemplate.BlazorServer)]
     [InlineData(DotNetTemplate.ClassLib)]
